test: move tests for getPartials to generate-html
